added rain text to measurand text hierarchy
