+++ lib/wrf_forecast/forecast/forecast_text.rb
@@ -2,13 +2,12 @@
 # @Author: Benjamin Held
 # @Date:   2020-03-22 10:46:55
 # @Last Modified by:   Benjamin Held
-# @Last Modified time: 2020-05-06 18:47:41
+# @Last Modified time: 2020-05-06 18:57:39
 
 require 'wrf_library/wrf'
 require 'wrf_forecast/data/forecast_repository'
 require 'wrf_forecast/threshold'
 require 'wrf_forecast/text'
-require 'wrf_forecast/forecast/english/rain/rain_text'
 
 module WrfForecast
 
@@ -101,7 +100,7 @@ module WrfForecast
       extreme_values = repository.extreme_values[:rain]
       hourly_rain = repository.hourly_rain
       threshold = handler.rain_threshold.indicators
-      @rain_text = WrfForecast::RainText.new(extreme_values, hourly_rain, threshold)
+      @rain_text = WrfForecast::Text::RainText.new(extreme_values, hourly_rain, threshold)
       nil
     end
