+++ lib/wrf_forecast/text.rb
@@ -2,12 +2,13 @@
 # @Author: Benjamin Held
 # @Date:   2020-05-06 18:11:59
 # @Last Modified by:   Benjamin Held
-# @Last Modified time: 2020-05-06 18:47:28
+# @Last Modified time: 2020-05-06 18:58:53
 
 module WrfForecast
 
   require "wrf_forecast/forecast/measurand_text"
   require "wrf_forecast/forecast/english/temperature_text"
   require "wrf_forecast/forecast/english/wind_text"
+  require "wrf_forecast/forecast/english/rain_text"
 
 end
